split up rendering functions
